\name{is.discdd.misclass}
\alias{is.discdd.misclass}

\title{
Class \code{discdd.misclass}
}
\description{
Tests if its argument is an object of class \code{discdd.misclass} (see Details 
of the function \link{discdd.misclass}). 
}
\usage{
is.discdd.misclass(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{discdd.misclass}, and \code{FALSE} otherwise.
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
    \link{discdd.misclass}.
}
