\name{is.discdd.predict}
\alias{is.discdd.predict}

\title{
Class \code{discdd.predict}
}
\description{
Tests if its argument is an object of class \code{discdd.predict} (see Details 
of the function \link{discdd.predict}). 
}
\usage{
is.discdd.predict(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{discdd.predict}, and \code{FALSE} otherwise.
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
    \link{discdd.predict}.
}
