\name{is.dstatis}
\alias{is.dstatis}

\title{
Class \code{dstatis}
}
\description{
Tests if its argument is an object of class \code{dstatis} (see Details 
of the function \link{dstatis.inter}).
}
\usage{
is.dstatis(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{dstatis}, and \code{FALSE} otherwise. 
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
    \link{dstatis.inter}.
}
