\name{is.fdiscd.predict}
\alias{is.fdiscd.predict}

\title{
Class \code{fdiscd.predict}
}
\description{
Tests if its argument is an object of class \code{fdiscd.predict} (see Details 
of the function \link{fdiscd.predict})..
}
\usage{
is.fdiscd.predict(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{fdiscd.predict}, and \code{FALSE} otherwise. 
}
\author{
    Rachid Boumaza, Pierre Santagostini, Smail Yousfi, Sabine Demotes-Mainard.
}

\seealso{
    \link{fdiscd.predict}.
}
