\name{is.fhclustd}
\alias{is.fhclustd}

\title{
Class \code{fhclustd}
}
\description{
Tests if its argument is an object of class \code{fhclustd} (see Details 
of the function \link{fhclustd}).
}
\usage{
is.fhclustd(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{fhclustd}, and \code{FALSE} otherwise. 
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
    \link{fhclustd}.
}
