\name{is.fmdsd}
\alias{is.fmdsd}

\title{
Class \code{fmdsd}
}
\description{
Tests if its argument is an object of class \code{fmdsd} (see Details 
of the function \link{fmdsd}).
}
\usage{
is.fmdsd(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{fmdsd}, and \code{FALSE} otherwise.
}
\author{
    Rachid Boumaza, Pierre Santagostini, Smail Yousfi, Sabine Demotes-Mainard.
}

\seealso{
    \link{fmdsd}.
}
