\name{is.folderh}
\alias{is.folderh}
\title{
Class \code{folderh}
}
\description{
Tests if its argument is an object of class \code{folderh} (see \code{\link{folderh}}).
}
\usage{
is.folderh(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{folderh}, and \code{FALSE} otherwise. 
}
\author{
Rachid Boumaza, Pierre Santagostini, Smail Yousfi, Sabine Demotes-Mainard
}

\seealso{
\code{\link{folderh}} to create an object of class \code{folderh}.
}
