\name{is.folder}
\alias{is.folder}
\title{
Class \code{folder}
}
\description{
Tests if its argument is an object of class \code{folder} (see \code{\link{folder}}).
}
\usage{
is.folder(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{folder}, and \code{FALSE} otherwise. 
}
\author{
Rachid Boumaza, Pierre Santagostini, Smail Yousfi, Sabine Demotes-Mainard
}

\seealso{
\code{\link{folder}} to create an object of class \code{folder}.
}
