\name{is.foldermtg}
\alias{is.foldermtg}
\title{
Class \code{foldermtg}
}
\description{
Tests if its argument is an object of class \code{foldermtg} (see \code{\link{read.mtg}}).
}
\usage{
is.foldermtg(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{foldermtg}, and \code{FALSE} otherwise. 
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
\code{\link{read.mtg}} to read a MTG file and create an object of class \code{foldermtg}.
}
