\name{is.foldert}
\alias{is.foldert}
\title{
Class \code{foldert}
}
\description{
Tests if its argument is an object of class \code{foldert} (see \code{\link{foldert}}).
}
\usage{
is.foldert(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{foldert}, and \code{FALSE} otherwise. 
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
\code{\link{foldert}} to create an object of class \code{foldert}.
}
