\name{is.fpcad}
\alias{is.fpcad}

\title{
Class \code{fpcad}
}
\description{
Tests if its argument is an object of class \code{fpcad} (see Details 
of the function \link{fpcad}).
}
\usage{
is.fpcad(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{fpcad}, and \code{FALSE} otherwise. 
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
    \link{fpcad}.
}
