\name{is.mdsdd}
\alias{is.mdsdd}

\title{
Class \code{mdsdd}
}
\description{
Tests if its argument is an object of class \code{mdsdd} (see Details 
of the function \link{mdsdd}).
}
\usage{
is.mdsdd(x)
}
\arguments{
  \item{x}{
       object to be tested.
}
}
\value{
    \code{TRUE} if its argument is of class \code{mdsdd}, and \code{FALSE} otherwise.
}
\author{
Rachid Boumaza,  Pierre Santagostini, Smail Yousfi, Gilles Hunault, Sabine Demotes-Mainard
}

\seealso{
    \link{mdsdd}.
}
